backend/requirements-dev.txt
# Local Development Requirements
# These are only needed for training models locally
# DO NOT use in production deployment

Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.21
python-dotenv==1.0.0

# AI/ML Dependencies - needed for local training only
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.2.2
joblib==1.3.2

# Development tools
gunicorn==21.2.0
psycopg2-binary==2.9.9
pytest==8.3.5
